# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    The user is presented with a list of conversations with other users or with groups, sorted in reverse
    chronological order. Each element in the list must display the username of the other person or the
    group name, the user profile photo or the group photo, the date and time of the latest message,
    the preview (snippet) of the text message, or an icon for a photo message. The user can start a new
    conversation with any other user of WASAText, and this conversation will automatically be added
    to the list. The user can search for other users via the username and see all the existing WASAText
    usernames.
    The user can create a new group with any number of other WASAText users to start a conversation.
    Group members can add other users to the group, but users cannot join groups on their own or even
    see groups they aren't a part of. Additionally, users have the option to leave a group at any time.
    The user can open a conversation to view all exchanged messages, displayed in reverse chronological
    order. Each message includes the timestamp, the content (whether text or photo), and the sender's
    username for received messages, or one/two checkmarks to indicate the status of sent messages. Any
    reactions (comments) on messages are also displayed, along with the names of the users who posted
    them.
...
    reactions at any time (a.k.a. uncomment).
...
  contact:
...
paths:
...
  /chats/{chat_id}/messages:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    post:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: send a message in the chat
      description: |-
        The user performing the action send a message in the chat
        associated with chat_id
      operationId: sendMessage
      requestBody:
        description: |-
          message to send
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MessageToSend"}
      responses:
        "201":
          description: Message send succesfully
          content:
            application/json:
              schema: 
                type: object
                description: id of the new message created
                properties: 
                  messageid: {$ref: "#/components/schemas/MessageId"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /chats/{chat_id}/forwardedmessages:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    post:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: forward a message in the chat
      description: |-
        The user performing the action forward a message in the chat
        associated with chat_id
      operationId: forwardMessage
      requestBody:
        description: |-
          message to forward
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MessageId"}
      responses:
        "201":
          description: Message forwarded succesfully
          content:
            application/json:
              schema:
                type: object
                description: id of the new message created
                properties: 
                  messageid: {$ref: "#/components/schemas/MessageId"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /chats/{chat_id}/messages/{message_id}:
    parameters:
    - $ref: "#/components/parameters/ChatId"
    - $ref: "#/components/parameters/MessageId"
    delete:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: delete a message in the chat
      description: |-
        The user performing the action delete the message associated to message_id 
        in the chat associated with chat_id
      operationId: deleteMessage
      responses:
        "204":
          description: Message deleted successfully
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
components:
  parameters:
    ChatId:
      name: chat_id
      in: path
      required: true
      description: Chat ID
      schema:
        type: integer
        example: 1
        readOnly: true
...
    MessageId:
      name: message_id
      in: path
      required: true
      description: Message ID
      schema:
        type: integer
        example: 1
        readOnly: true
...
  schemas:
...
    Photo:
      title: photo
      description: byte data of the photo
      type: string
      example: "data:image/jpg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/AAAAAAAAAAAAAQIDAAQF/"
      pattern: '.*'
      minLength: 1
      maxLength: 999
...
    MessageId:
      title: messageid
      description: message id
      type: integer
      example: 33
...
    TextData:
      title: textdata
      description: Data of a text or a photo message
      type: string
      example: "Ciao"
      pattern: '^(.|\n)*$'
      minLength: 1
      maxLength: 999
...
    MessageToSend:
      title: messagetosend
      description: info to create a new message
      type: object
      properties:
        text: {$ref: "#/components/schemas/TextData"}
        photo: {$ref: "#/components/schemas/Photo"}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    Unauthorized:
      description: Access token is missing or invalid.
    NotFound:
      description: The server cannot find the requested resource.
    InternalServerError:
      description: The server encounted an internal error. Further info in server logs.